\name{tuneRF-package}
\alias{tuneRF-package}
\alias{tuneRF}
\docType{package}
\title{
\packageTitle{tuneRF}
}
\description{
\packageDescription{tuneRF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tuneRF}
\packageIndices{tuneRF}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{tuneRF}

Maintainer: \packageMaintainer{tuneRF}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
